zng_view_angle

Function prefer_angle_egl

pub fn prefer_angle_egl(
    child: impl IntoUiNode,
    enable: impl IntoValue<bool>,
) -> UiNode

P Signal the view-process to try and use EGL for this window when the ANGLE DLLs can be found and the render mode is not software.

Note that the libEGL.dll and libGLESv2.dll files must be distributed with the executable.

Note that this will enable ANGLE just for this window, use register_root_extender to enable angle for all windows.

This does nothing when app is not running on Windows.